[NDK][NTOS] - Move KeProfileInterrupt[WithSource] declarations to NDK since they're exported - NtCreateProfile's fourth parameter is SIZE_T - Add NtCreateProfileEx prototype CORE-10066
svn path=/trunk/; revision=68861

+++ reactos/ntoskrnl/ex/profile.c
@@ -95,7 +95,7 @@ NTAPI
 NtCreateProfile(OUT PHANDLE ProfileHandle,
                 IN HANDLE Process OPTIONAL,
                 IN PVOID RangeBase,
-                IN ULONG RangeSize,
+                IN SIZE_T RangeSize,
                 IN ULONG BucketSize,
                 IN PVOID Buffer,
                 IN ULONG BufferSize,
